% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviz.R
\name{bnviz}
\alias{bnviz}
\title{Bayesian Network visuallization.}
\usage{
bnviz()
}
\description{
bnviz is a Shiny web application for Bayesian Network visualization
}
\examples{
if (interactive()) {
  bnviz()
}
}
